fix: rewrite about component
