+++ src/pages/index.tsx
@@ -1,7 +1,8 @@
+import About from '@/components/about'
 import React from 'react'
 
 const Landing = () => {
-	return <>Land</>
+	return <About />
 }
 
 export default Landing
